code-samples/Python/semantic-kernel-multi-agent/requirements.txt
openai
semantic-kernel
python-dotenv